添加需求 part 1

+++ src/main/java/io/github/weightrack/controller/CoalTypeController.java
@@ -1,12 +1,17 @@
 package io.github.weightrack.controller;
 
 import com.alibaba.fastjson2.JSON;
-import io.github.weightrack.service.CoalTypeService;
 import io.github.weightrack.module.CoalType;
+import io.github.weightrack.service.CoalTypeService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.List;
 
@@ -19,12 +24,19 @@ public class CoalTypeController {
     @ResponseBody
     @GetMapping("/coalType")
     public String coalType() {
-        CoalType[] coalTypes = coalTypeService.getCoalType();
+        CoalType[] coalTypes = coalTypeService.getCoalTypes();
         return JSON.toJSONString(coalTypes);
     }
 
     @PostMapping("/coalType/delete")
-    public String deleteCoalType(@RequestParam List<String> coalTypes) {
+    public String deleteCoalType(@RequestParam(value = "coalTypes", required = false) List<String> coalTypes, Model model) {
+        if (coalTypes == null) {
+            model.addAttribute("error", "请选择要删除的煤种");
+            model.addAttribute("coalTypes", coalTypeService.getCoalTypes());
+            return "coalTypeEdit";
+        }
+        Logger logger = LoggerFactory.getLogger(CoalTypeController.class);
+        logger.info("删除煤种：{}", coalTypes);
         for (String coalType : coalTypes) {
             coalTypeService.deleteCoalTypeByName(coalType);
         }
@@ -34,7 +46,7 @@ public class CoalTypeController {
 
     @GetMapping("/coalTypeEdit")
     public String coalTypeEdit(Model model) {
-        model.addAttribute("coalTypes", coalTypeService.getCoalType());
+        model.addAttribute("coalTypes", coalTypeService.getCoalTypes());
         return "coalTypeEdit";
     }
 }
